# docs/swagger.yaml
basePath: /
definitions:
...
  serializer.TireJSON:
    properties:
      description:
        type: string
      is_delete:
        type: boolean
      photo:
        type: string
      tire_id:
        type: integer
      tire_material_coefficient:
        type: number
      tire_thickness_coefficient:
        type: number
      tire_title:
        type: string
      video:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для расчёта давления в шинах
  title: Tire Pressure API
  version: "1.0"
paths:
...
  /api/tires:
    get:
      parameters:
      - description: Поиск по названию шины
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serializer.TireJSON'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список шин
      tags:
      - tires
    post:
      consumes:
      - multipart/form-data
      description: multipart/form-data или application/json; только для модераторов
      parameters:
      - description: Название шины
        in: formData
        name: tire_title
        type: string
      - description: Коэффициент материала
        in: formData
        name: tire_material_coefficient
        type: number
      - description: Коэффициент толщины
        in: formData
        name: tire_thickness_coefficient
        type: number
      - description: Описание
        in: formData
        name: description
        type: string
      - description: Фото шины
        in: formData
        name: photo
        type: file
      - description: Видео шины
        in: formData
        name: video
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/serializer.TireJSON'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Требуется роль модератора
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать шину
      tags:
      - tires
  /api/tires/{id}:
    get:
      parameters:
      - description: ID шины
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.TireJSON'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Шина по ID
      tags:
      - tires
...
  /tire/{id}:
    get:
      parameters:
      - description: ID шины
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML-страница
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Страница шины по ID
      tags:
      - tires-html
  /tires:
    get:
      parameters:
      - description: Поиск по названию шины
        in: query
        name: query
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML-страница
          schema:
            type: string
      summary: Главная страница (список шин)
      tags:
      - tires-html
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"